server/requirements.txt
django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
gunicorn==21.2.0
psycopg[binary]==3.2.3
python-dotenv==1.0.1
dj-database-url==2.3.0

